Methods with generic return type and annotation
Methods that define a return type and an annotation on the return type fail to parse. This changeset  adds support for this type of methods.

I didn't find a way to put the annotation on the TypeDef object that represents the return type bu that's OK for my current use case.